requirements.txt
aiohttp
aiosqlite
nextcord
python-dotenv
pytz
requests
translate
psutil
PyGithub